Changes in about us page
